requirements.txt
asgiref==3.7.2
astunparse==1.6.3
black==23.11.0
click==8.1.7
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.7
django-click==2.3.0
djangorestframework==3.14.0
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
gunicorn==21.2.0
html-json-forms==1.1.1
mypy-extensions==1.0.0
natural-keys==2.1.0
openpyxl==3.0.9
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.0.0
psycopg2-binary==2.9.9
pystache==0.6.5
pytz==2023.3.post1
pyxform==1.12.2
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
whitenoise==6.6.0
wq==2.0.0
wq.app==2.0.0
wq.build==2.0.0
wq.create==2.0.0
wq.db==2.0.0
xlrd==2.0.1
xlsconv==2.0.0
